requirements-anonimizador.txt
aiopulsar-py==0.1.0a0
fastapi==0.89.1
uvicorn[standard]==0.20.0
pulsar-client==3.6.0
pulsar-client[avro]==3.6.0
SQLAlchemy==1.4.44
requests==2.28.2